% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_components.R
\name{get_api_components}
\alias{get_api_components}
\title{Get a table of the NSSP-ESSENCE API Components}
\usage{
get_api_components(url)
}
\arguments{
\item{url}{NSSP-ESSENCE API URL. Intended to be used in other \code{synchron} functions, but can be useful on its own as well}
}
\value{
a data frame with two columns, \code{component} (i.e., query or API component) and \code{value} (i.e, query criteria)
}
\description{
Get a table of the NSSP-ESSENCE API Components
}
\examples{
url <- "https://essence2.syndromicsurveillance.org/nssp_essence/api/
dataDetails?datasource=va_hosp&startDate=1Jan2023&medicalGroupingSystem
=essencesyndromes&userId=5099&endDate=31Dec2024&cDeath=no&percentParam
=noPercent&hospFacilityType=emergency\%20care&aqtTarget=DataDetails
&ccddCategory=cdc\%20motor\%20vehicle\%20crash\%20occupant\%20injury\%20v1
&geographySystem=hospitalregion&detector=probrepswitch&timeResolution=daily
&hasBeenE=1"
url_comps <- get_api_components(url)
}
