% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_essence_data.R
\name{get_essence_data}
\alias{get_essence_data}
\title{Return output of an NSSP-ESSENCE API URL}
\usage{
get_essence_data(url, start_date = NULL, end_date = NULL)
}
\arguments{
\item{url}{NSSP-ESSENCE URL}

\item{start_date}{use format "YYYY-MM-DD"}

\item{end_date}{use format "YYY-MM-DD"}
}
\value{
dependent on API type
}
\description{
\code{get_essence_data} is a wrapper function that uses the \code{myProfile} credentials
and API pull functions from the \code{Rnssp} package. It detects what the API type
is and returns the object associated with that API type. Developed based on
examples from \url{https://cdn.ymaws.com/www.cste.org/resource/resmgr/RStudio_ESSENCE_API_Guide_J.html}
}
\examples{
ts_graph_url <- "https://essence2.syndromicsurveillance.org/nssp_essence/api
/timeSeries/graph?datasource=va_hosp&startDate=4Oct2024&medicalGroupingSystem
=essencesyndromes&endDate=2Jan2025&percentParam=noPercent&aqtTarget=TimeSeries
&ccddCategory=cdc\%20respiratory\%20syncytial\%20virus\%20dd\%20v1&geographySystem
=hospitalregion&detector=nodetectordetector&timeResolution=daily&hasBeenE=1"
try(get_essence_data(ts_graph_url))

data_details_json_url <- "https://essence2.syndromicsurveillance.org/
nssp_essence/api/dataDetails?datasource=va_hosp&startDate=04Oct24
&medicalGroupingSystem=essencesyndromes&endDate=02Jan25&percentParam=noPercent
&aqtTarget=TimeSeries&ccddCategory=cdc\%20respiratory\%20syncytial\%20virus\%20dd\%20v1
&geographySystem=hospitalregion&detector=nodetectordetector&
timeResolution=daily&hasBeenE=1"
try(get_essence_data(data_details_json_url))

table_builder_json_url <- "https://essence2.syndromicsurveillance.org/
nssp_essence/api/tableBuilder?datasource=va_hosp&startDate=4Oct2024
&medicalGroupingSystem=essencesyndromes&endDate=2Jan2025&percentParam=noPercent
&aqtTarget=TableBuilder&ccddCategory=cdc\%20respiratory\%20syncytial\%20virus\%20dd\%20v1
&geographySystem=hospitalregion&detector=nodetectordetector&timeResolution=daily
&hasBeenE=1&rowFields=timeResolution&columnField=ccddCategory"
try(get_essence_data(table_builder_json_url))
}
