% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dd_varname_crosswalk.R
\docType{data}
\name{query_dd_varname_crosswalk}
\alias{query_dd_varname_crosswalk}
\title{Query API and Data Details API Crosswalk R Object}
\format{
\subsection{\code{query_dd_varname_crosswalk}}{

A data frame with 36 rows and 2 columns:
\describe{
\item{QueryAPI}{Name used when querying the field}
\item{DataDetailsAPI}{Name used in data details output}
}
}
}
\source{
\url{https://www.cdc.gov/nssp/biosense/docs/NSSP-Data-Dictionary-508.xlsx}
}
\usage{
query_dd_varname_crosswalk
}
\description{
A crosswalk of the API fields used to query a field and their corresponding data details fields
}
\keyword{datasets}
